Extract OverviewTree to its own file and add unit tests
